% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{accurate_align}
\alias{accurate_align}
\title{peak.table}
\usage{
accurate_align(peak.table, simple.data, use.int.tol)
}
\arguments{
\item{peak.table}{peak.table}

\item{simple.data}{simple.data}

\item{use.int.tol}{use.int.tol}
}
\value{
result
}
\description{
peak.table
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
