% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{align_2batch}
\alias{align_2batch}
\title{align_2batch}
\usage{
align_2batch(
  batch1,
  batch2,
  mz.error.sd,
  rt.error.sd,
  int.error.sd,
  fold = 4,
  mz.weight = 0.4,
  rt.weight = 0.4,
  int.weight = 0.2
)
}
\arguments{
\item{batch1}{batch1}

\item{batch2}{batch2}

\item{mz.error.sd}{mz.error.sd}

\item{rt.error.sd}{rt.error.sd}

\item{int.error.sd}{int.error.sd}

\item{fold}{fold}

\item{mz.weight}{mz.weight}

\item{rt.weight}{rt.weight}

\item{int.weight}{int.weight}
}
\value{
result
}
\description{
peak.table
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
