% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{alignBatch}
\alias{alignBatch}
\title{alignBatch}
\usage{
alignBatch(
  object,
  combine.mz.tol = 25,
  combine.rt.tol = 30,
  use.int.tol = FALSE
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{combine.mz.tol}{m/z tolerance for batch alignment, default is 25 ppm.}

\item{combine.rt.tol}{RT tolerance for batch alignment, default is 30 seconds.}

\item{use.int.tol}{Whether use intensity match for batch aglignment.}
}
\value{
A new metflowClass object.
}
\description{
Align different batch peaks tables.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
