% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{baINTplot}
\alias{baINTplot}
\title{baINTplot}
\usage{
baINTplot(simple.data)
}
\arguments{
\item{simple.data}{simple.data}
}
\value{
result
}
\description{
baINTplot
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
