% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{baMZplot}
\alias{baMZplot}
\title{baMZplot}
\usage{
baMZplot(simple.data)
}
\arguments{
\item{simple.data}{data}
}
\value{
result
}
\description{
baMZplot
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
