% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{baRTplot}
\alias{baRTplot}
\title{baRTplot}
\usage{
baRTplot(simple.data)
}
\arguments{
\item{simple.data}{simple.data}
}
\value{
result
}
\description{
baRTplot
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
