% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_for_data_integration}
\alias{check_for_data_integration}
\title{check_for_data_integration}
\usage{
check_for_data_integration(
  object,
  method = c("qc_mean", "qc_median", "subject_mean", "subject_median")
)
}
\arguments{
\item{object}{mass_dataset}

\item{method}{method for integrate_data}
}
\value{
error information
}
\description{
check_for_data_integration
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
