% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_for_qc_normalization}
\alias{check_for_qc_normalization}
\title{check_for_qc_normalization}
\usage{
check_for_qc_normalization(object)
}
\arguments{
\item{object}{mass_dataset}
}
\value{
error information
}
\description{
check_for_qc_normalization
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
