% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{cvMSE}
\alias{cvMSE}
\title{cvMSE}
\usage{
cvMSE(qc, qc.order, begin1, end1, step1)
}
\arguments{
\item{qc}{qc}

\item{qc.order}{qc.order}

\item{begin1}{begin1}

\item{end1}{end1}

\item{step1}{step1}
}
\value{
result
}
\description{
cvMSE
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
