% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outlier_samples.R
\name{detect_outlier}
\alias{detect_outlier}
\title{Detect outlier samples}
\usage{
detect_outlier(
  object,
  na_percentage_cutoff = 0.5,
  sd_fold_change = 6,
  mad_fold_change = 6,
  dist_p_cutoff = 0.05
)
}
\arguments{
\item{object}{A mass_dataset object.}

\item{na_percentage_cutoff}{na_percentage_cutoff}

\item{sd_fold_change}{sd_fold_change}

\item{mad_fold_change}{mad_fold_change}

\item{dist_p_cutoff}{dist_p_cutoff}
}
\value{
A new mass_dataset object.
}
\description{
Detect outlier samples
}
\examples{
library(massdataset)
data("expression_data")
data("sample_info")
data("variable_info")
object =
  create_mass_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

object =
  object \%>\%
  log() \%>\%
  scale()

outlier_samples =
  object \%>\%
  detect_outlier()

extract_outlier_table(outlier_samples)

###MV plot
massdataset::show_sample_missing_values(object = object,
                                        color_by = "class",
                                        percentage = TRUE)
}
\seealso{
https://privefl.github.io/blog/detecting-outlier-samples-in-pca/
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
