% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outlier_samples.R
\name{extract_outlier_table}
\alias{extract_outlier_table}
\title{extract_outlier_table}
\usage{
extract_outlier_table(object)
}
\arguments{
\item{object}{outlier_samples class object.}
}
\value{
A data frame.
}
\description{
extract_outlier_table
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
