% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_rsd}
\alias{filter_rsd}
\title{filter_rsd}
\usage{
filter_rsd(object, slot = c("Subject", "QC"), rsd.cutoff = 30)
}
\arguments{
\item{object}{A metflowClass object.}

\item{slot}{What group samples are used to calculate RSD.}

\item{rsd.cutoff}{RSD cutoff.}
}
\value{
A new metflowClass object.
}
\description{
Filter peaks according to RSD.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
