% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter samples}
\usage{
filter_samples(object, min.fraction.peak = 0.8)
}
\arguments{
\item{object}{A metflowClass object.}

\item{min.fraction.peak}{Peaks minimun fraction in subject samples.}
}
\value{
A new metflowClass object.
}
\description{
Filter samples.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
