% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filterPeak}
\alias{filterPeak}
\title{filterPeak}
\usage{
filterPeak(
  object,
  min.fraction.qc = 0.8,
  min.fraction = 0.8,
  min.subject.qc.ratio = 2,
  dl.qc.r2.cutoff = 0.7,
  silence.deprecated = FALSE
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{min.fraction.qc}{Peaks minimum fraction in QC samples.}

\item{min.fraction}{Peaks minimun fraction in subject samples.}

\item{min.subject.qc.ratio}{Peak intensity ratio in subject and blank samples.}

\item{dl.qc.r2.cutoff}{R2 cutoff for dilution QC.}

\item{silence.deprecated}{Silence deprecated information or not.}
}
\value{
A new metflowClass object.
}
\description{
Filter peaks.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
