% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filterRSD}
\alias{filterRSD}
\title{filterRSD}
\usage{
filterRSD(
  object,
  slot = c("Subject", "QC"),
  rsd.cutoff = 30,
  silence.deprecated = FALSE
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{slot}{What group samples are used to calculate RSD.}

\item{rsd.cutoff}{RSD cutoff.}

\item{silence.deprecated}{Silence deprecated information or not.}
}
\value{
A new metflowClass object.
}
\description{
Filter peaks according to RSD.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
