% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filterSample}
\alias{filterSample}
\title{filterSample}
\usage{
filterSample(object, min.fraction.peak = 0.8, silence.deprecated = FALSE)
}
\arguments{
\item{object}{A metflowClass object.}

\item{min.fraction.peak}{Peaks minimun fraction in subject samples.}

\item{silence.deprecated}{Silence deprecated information or not.}
}
\value{
A new metflowClass object.
}
\description{
Filter samples.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
