% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masscleaner_logo.R
\name{get_masscleaner_version}
\alias{get_masscleaner_version}
\title{get_masscleaner_version}
\usage{
get_masscleaner_version()
}
\value{
masscleaner_version
}
\description{
get_masscleaner_version
}
\examples{
get_masscleaner_version
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
