% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4dataNormalization.R
\name{get_peak_int_distribution}
\alias{get_peak_int_distribution}
\title{Get peak intensity distributation plot}
\usage{
get_peak_int_distribution(object, peak_name, interactive = TRUE)
}
\arguments{
\item{object}{A metflowClass object.}

\item{peak_name}{Peak name.}

\item{interactive}{Interactive plot or not.}
}
\value{
A ggplot2 object.
}
\description{
Get peak intensity distributation plot.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
