% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{getBatchAlignmentInfo}
\alias{getBatchAlignmentInfo}
\title{getBatchAlignmentInfo}
\usage{
getBatchAlignmentInfo(raw.data, rough.align.data, accurate.align.data)
}
\arguments{
\item{raw.data}{raw.data}

\item{rough.align.data}{rough.align.data}

\item{accurate.align.data}{accurate.align.data}
}
\value{
result
}
\description{
getBatchAlignmentInfo
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
