% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mv.R
\name{impute_mv}
\alias{impute_mv}
\title{Impute MV in data.}
\usage{
impute_mv(
  object,
  method = c("knn", "rf", "mean", "median", "zero", "minimum", "bpca", "svdImpute",
    "ppca"),
  k = 10,
  rowmax = 0.5,
  colmax = 0.8,
  maxp = 1500,
  rng.seed = 362436069,
  maxiter = 10,
  ntree = 100,
  decreasing = FALSE,
  nPcs = 2,
  maxSteps = 100,
  threshold = 1e-04,
  ...
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{method}{Imputation method. It
contains "knn", "rf" (missForest), "mean", "median", "zero", "minium",
"bpca" (BPCA), "svd" (SVD) and "ppca" (PPCA). Default is "knn".
The detial of
this method can be find in detail and reference paperes.}

\item{k}{See ?impute.knn}

\item{rowmax}{See ?impute.knn}

\item{colmax}{See ?impute.knn}

\item{maxp}{See ?impute.knn}

\item{rng.seed}{See ?impute.knn}

\item{maxiter}{See ?missForest}

\item{ntree}{See ?missForest}

\item{decreasing}{See ?missForest}

\item{nPcs}{See ?bpca}

\item{maxSteps}{See ?bpca}

\item{threshold}{See ?bpca}

\item{...}{Other arguments.}
}
\value{
A new metflowClass object.
}
\description{
Impute MV in data.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
