% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_data.R
\name{integrate_data}
\alias{integrate_data}
\title{integrate_data}
\usage{
integrate_data(
  object,
  method = c("qc_mean", "qc_median", "subject_mean", "subject_median")
)
}
\arguments{
\item{object}{object.}

\item{method}{Which method do you want to use?
subject_mean or qc_mean, default is qc_mean.}
}
\value{
Return a mass_dataset which has been integrated.
}
\description{
Integrate different batch datasets together.
}
\details{
Data integration is a necessary step for multiple batch dataset.
\url{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:abe41368-d08d-4806-871f-3aa035d21743}{Dunn's}
method has been used in this function.
}
\examples{
\dontrun{
data("object1", package = "demodata")
data("object2", package = "demodata")

object =
align_batch(x = object1, y = object2, return_index = FALSE)

object = impute_mv(object = object, method = "zero")

get_mv_number(object)

new_object =
integrate_data(object = object, method = "qc_mean")

massdataset::intensity_plot(object = object,
                            variable_index = 4,
                            order_by = "injection.order",color_by = "batch")

massdataset::intensity_plot(object = new_object,
                            variable_index = 2,
                            order_by = "injection.order",color_by = "batch")
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
