% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{loessNor}
\alias{loessNor}
\title{loessNor}
\usage{
loessNor(
  subject_data,
  qc_data,
  subject_order,
  qc_order,
  optimization = TRUE,
  begin = 0.5,
  end = 1,
  step = 0.2,
  path = ".",
  threads = 4
)
}
\arguments{
\item{subject_data}{subject_data}

\item{qc_data}{qc_data}

\item{subject_order}{subject_order}

\item{qc_order}{qc_order}

\item{optimization}{optimization}

\item{begin}{begin}

\item{end}{end}

\item{step}{step}

\item{path}{path}

\item{threads}{threads}
}
\value{
result
}
\description{
loessNor
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
