% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{masscleaner_packages}
\alias{masscleaner_packages}
\title{List all packages in the masscleaner}
\usage{
masscleaner_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include masscleaner in the list?}
}
\value{
masscleaner_packages
}
\description{
List all packages in the masscleaner
}
\examples{
masscleaner_packages()
}
