% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{matchScore}
\alias{matchScore}
\title{matchScore}
\usage{
matchScore(error, sd)
}
\arguments{
\item{error}{error}

\item{sd}{sd}
}
\value{
result
}
\description{
matchScore
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
