% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{MRImatch}
\alias{MRImatch}
\title{MRImatch}
\usage{
MRImatch(
  data1,
  data2,
  mz.tol,
  rt.tol = 30,
  rt.error.type = c("relative", "abs"),
  int.tol = 1
)
}
\arguments{
\item{data1}{data1}

\item{data2}{data2}

\item{mz.tol}{mz.tol}

\item{rt.tol}{rt.tol}

\item{rt.error.type}{rt.error.type}

\item{int.tol}{int.tol}
}
\value{
result
}
\description{
MRImatch
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
