% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_scale}
\alias{new_scale}
\title{new_scale}
\usage{
new_scale(new_aes)
}
\arguments{
\item{new_aes}{new_aes}
}
\value{
result
}
\description{
new_scale
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
