% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{normalize_data}
\usage{
normalize_data(
  object,
  method = c("svr", "total", "median", "mean", "pqn", "loess"),
  keep.scale = TRUE,
  begin = 0.5,
  end = 1,
  step = 0.2,
  threads = 4
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{method}{Normalization method, mean, median, total svr or loess,
default is svr. Please see the details.}

\item{keep.scale}{Remain scale or not. Default is TRUE.}

\item{begin}{0.5}

\item{end}{1}

\item{step}{0.2}

\item{threads}{4}
}
\value{
A new metflowClass object.
}
\description{
Data normalization.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
