% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data_loess.R
\name{normalize_data_loess}
\alias{normalize_data_loess}
\title{normalize_data_loess}
\usage{
normalize_data_loess(
  subject_data,
  qc_data,
  subject_order,
  qc_order,
  optimization = TRUE,
  begin = 0.5,
  end = 1,
  step = 0.2,
  threads = 4
)
}
\arguments{
\item{subject_data}{subject_data row is variable, column is sample}

\item{qc_data}{subject_data row is variable, column is sample}

\item{subject_order}{subject_order}

\item{qc_order}{qc_order}

\item{optimization}{optimization}

\item{begin}{begin see ?loess}

\item{end}{end ?loess}

\item{step}{step ?loess}

\item{threads}{threads}
}
\value{
result
}
\description{
normalize_data_loess
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
