% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data_sample_wise.R
\name{normalize_data_median}
\alias{normalize_data_median}
\title{Normalize data using total.}
\usage{
normalize_data_median(x, keep_scale = TRUE)
}
\arguments{
\item{x}{expression data, row is variables, column is sample.}

\item{keep_scale}{keep scale or not.}
}
\value{
Normalized expression data.
}
\description{
Normalize data using total.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
