% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data_pqn.R
\name{normalize_data_pqn}
\alias{normalize_data_pqn}
\title{Perform Probabilistic Quotient Normalization}
\usage{
normalize_data_pqn(x, pqn_reference = "median", pgn_reference_sample = NULL)
}
\arguments{
\item{x}{matrix to normalize. row is variable and column is sample.}

\item{pqn_reference}{normalization reference: "mean" for using the overall
average of variables as reference
or "median" (default) for using the overall median of variables as reference}

\item{pgn_reference_sample}{vector of index to specify
samples which average to use as reference}
}
\value{
Normalized expression dataset.
}
\description{
Perform Probabilistic Quotient Normalization
}
\details{
First a total area normalization should be done before PQN is applied.
}
\references{
Dieterle, F., Ross, A., Schlotterbeck, G. & Senn, H. Probabilistic Quotient
Normalization as Robust Method to Account for Dilution of Complex Biological Mixtures.
Application in H1 NMR Metabonomics. Anal. Chem. 78, 4281-4290 (2006).
}
\author{
E. Nevedomskaya

Rico Derks
}
