% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data_svr.R
\name{normalize_data_svr}
\alias{normalize_data_svr}
\title{svrNor}
\usage{
normalize_data_svr(
  subject_data,
  qc_data,
  subject_order,
  qc_order,
  multiple = 5,
  threads = 3
)
}
\arguments{
\item{subject_data}{subject_data.}

\item{qc_data}{qc_data}

\item{subject_order}{subject_order}

\item{qc_order}{qc_order}

\item{multiple}{multiple}

\item{threads}{threads}
}
\value{
result
}
\description{
svrNor
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
