% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data_loess.R
\name{optimize_loess_span}
\alias{optimize_loess_span}
\title{optimize_loess_span}
\usage{
optimize_loess_span(
  x,
  y,
  degree_range = c(1, 2),
  span_range = seq(0.2, 0.6, 0.1)
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{degree_range}{numeric vector}

\item{span_range}{numeric vector}
}
\value{
optimization result
}
\description{
optimize_loess_span
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
