% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outlier_samples.R
\docType{class}
\name{outlier_samples}
\alias{outlier_samples}
\title{outlier_samples}
\description{
An S4 class that stores the outlier sample information.
}
\section{Slots}{

\describe{
\item{\code{outlier_samples_table}}{outlier_samples_table}

\item{\code{parameter}}{parameter}

\item{\code{parameter}}{parameter}

\item{\code{intermediate}}{intermediate}

\item{\code{ref}}{red}
}}

