% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{rough_align}
\alias{rough_align}
\title{rough_align}
\usage{
rough_align(peak.table, combine.mz.tol = 25, combine.rt.tol = 30)
}
\arguments{
\item{peak.table}{peak.table}

\item{combine.mz.tol}{combine.mz.tol}

\item{combine.rt.tol}{combine.rt.tol}
}
\value{
result
}
\description{
rough_align
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
