% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{simply_data}
\alias{simply_data}
\title{simply_data}
\usage{
simply_data(data, combine.mz.tol = 5, combine.rt.tol = 30)
}
\arguments{
\item{data}{data}

\item{combine.mz.tol}{combine.mz.tol}

\item{combine.rt.tol}{combine.rt.tol}
}
\value{
result
}
\description{
simply_data
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
