% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{simplyData}
\alias{simplyData}
\title{simplyData}
\usage{
simplyData(data, combine.mz.tol = 5, combine.rt.tol = 30)
}
\arguments{
\item{data}{data}

\item{combine.mz.tol}{combine.mz.tol}

\item{combine.rt.tol}{combine.rt.tol}
}
\value{
result
}
\description{
simplyData
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
