% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{svr_function}
\alias{svr_function}
\title{svr_function}
\usage{
svr_function(index, sample, qc, sample.order, qc.order, multiple)
}
\arguments{
\item{index}{index}

\item{sample}{sample}

\item{qc}{qc}

\item{sample.order}{sample.order}

\item{qc.order}{qc.order}

\item{multiple}{multiple}
}
\description{
svr_function
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
