% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{svrNor}
\alias{svrNor}
\title{svrNor}
\usage{
svrNor(
  sample,
  qc,
  sample.order,
  qc.order,
  multiple = 5,
  path = ".",
  dimension1 = TRUE,
  threads = 3
)
}
\arguments{
\item{sample}{sample.}

\item{qc}{qc}

\item{sample.order}{sample.order}

\item{qc.order}{qc.order}

\item{multiple}{multiple}

\item{path}{path}

\item{dimension1}{dimension1}

\item{threads}{threads}
}
\value{
result
}
\description{
svrNor
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
