% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_batch.R
\name{SXTMTmatch2}
\alias{SXTMTmatch2}
\title{SXTMTmatch2}
\usage{
SXTMTmatch2(
  data1,
  data2,
  mz.tol,
  rt.tol = 30,
  rt.error.type = c("relative", "abs")
)
}
\arguments{
\item{data1}{data1}

\item{data2}{data2}

\item{mz.tol}{mz.tol}

\item{rt.tol}{rt.tol}

\item{rt.error.type}{rt.error.type}
}
\value{
result
}
\description{
SXTMTmatch2
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
