% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{normalize_data}
\usage{
normalize_data(
  object,
  method = c("svr", "total", "median", "mean", "pqn", "loess"),
  keep_scale = TRUE,
  pqn_reference = c("median", "mean"),
  begin = 0.5,
  end = 1,
  step = 0.2,
  multiple = 1,
  threads = 4
)
}
\arguments{
\item{object}{A mass_dataset object.}

\item{method}{Normalization method, mean, median, total svr or loess,
default is svr. Please see the details.}

\item{keep_scale}{Remain scale or not. Default is TRUE.}

\item{pqn_reference}{for pqn method.}

\item{begin}{0.5}

\item{end}{1}

\item{step}{0.2}

\item{multiple}{multiple}

\item{threads}{4}
}
\value{
A new mass_dataset object.
}
\description{
Data normalization.
}
\examples{
\dontrun{
data("object1", package = "demodata")
object1 = impute_mv(object1, method = "minimum")
object_mean = normalize_data(object = object1, method = "mean")
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
